# main-webbase/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for user management.",
        "title": "Fiber MongoDB API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/event": {
            "get": {
                "description": "Retrieve all events that the current user can see. This endpoint returns events visible to the current authenticated user. You can optionally filter the events using query parameters: \\n - `q` (string): Search text that matches the event's topic or description (case-insensitive). \\n - `role` (string, comma-separated): Filter by user role or organization path. Each value can be: \\n - A position key, e.g., `Lecturer` (matches `postedas.position_key`, case-insensitive exact match). \\n - An organization path, e.g., `/fac/eng/com` (matches `org_of_content`). \\n - Subtree prefix is supported using `/*`, e.g., `/fac/eng/*` matches `/fac/eng/com` or `/fac/eng/math`. \\n If no filters are applied, all events that the user can see (based on visibility rules) are returned. \\n Visibility rules: \\n - `public`: visible to everyone. \\n - `org`: visible only to users whose organization is included in the audience. \\n - `draft`: visible only to organizers within the same organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get all visible events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search text in topic or description",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of roles or org paths to filter",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of visible events with their next upcoming schedule",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an event with optional image upload and schedule creation",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create a new event",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Event image file upload",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "NodeID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event topic",
                        "name": "topic",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Event description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Organization responsible for content (e.g. /fac/eng/com)",
                        "name": "org_of_content",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "active",
                            "draft",
                            "inactive"
                        ],
                        "type": "string",
                        "description": "Event status",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum participants",
                        "name": "max_participation",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether event has a form (true/false)",
                        "name": "have_form",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Organization path of the posting role",
                        "name": "postedAs.org_path",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Position key of the posting role",
                        "name": "postedAs.position_key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Visibility JSON string, e.g., {\\",
                        "name": "visibility",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Schedules JSON array, e.g., [{\\",
                        "name": "schedules",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.EventCreateResult"
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid input)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: Cannot post as this role",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/event/accept/{event_id}": {
            "patch": {
                "description": "Mark a pending event as active.",
                "tags": [
                    "events"
                ],
                "summary": "Approve a pending event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EventDetail"
                        }
                    },
                    "400": {
                        "description": "Invalid Event ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "pending event not found or already processed",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/event/participate/{event_id}": {
            "post": {
                "description": "Join an event directly if it does not require a form",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Participate in an event (no form required)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participation successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid body"
                }
            }
        },
        "dto.EventCreateResult": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/models.Event"
                },
                "form_id": {
                    "type": "string"
                },
                "organizer_count": {
                    "type": "integer"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EventSchedule"
                    }
                }
            }
        },
        "dto.EventDetail": {
            "type": "object",
            "properties": {
                "current_participation": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "have_form": {
                    "type": "boolean"
                },
                "max_participation": {
                    "type": "integer"
                },
                "orgpath": {
                    "type": "string"
                },
                "picture_url": {
                    "type": "string"
                },
                "posted_as": {
                    "$ref": "#/definitions/models.PostedAs"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EventSchedule"
                    }
                },
                "status": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                },
                "user_participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserParticipant"
                    }
                },
                "visibility": {
                    "$ref": "#/definitions/models.Visibility"
                }
            }
        },
...
    }
}